big improvements
previously we were not checking whether a registration did exist in our
database before commiting to write it. Now, we're fixed that. Also, we
would like to remove a magic number that was burried in our code, 9, in
a test concerning the school's quota in the competition. Also, add a
field in the student registration form where the professor can inform if
the student needs some help to participate in the competition.